Add camera session protocol
